Make the redirect work for standalone or container
We mount the websocket demo on its own context, /ws, so that it can be
deployed with all the others. Unfortunately, although the redirect works
fine in WF, the websocket connection fails, probably due to something
we're not tickling correctly, e.g. it may only be looking for annotated
Endpoints or some crap.