Enable HTTP/2 and SSL with Undertow key/trust stores
Stores are taken from Undertow examples. The web demo now creates two
listeners with the additional SSL port, 8443 by default. I'm not sure
this will work well for Heroku/OpenShift.

Note the use of string keys in web/-main, allowing the values to be
overridden on the command line. This is particularly useful when running
from the jar since the alpn jar needs to be in the bootclasspath.